Improve Code Quality: Check for code quality violations and standardized formatting throughout. Refactor Commands into separate class files
