--- src/main/java/EmptyTodo.java
@@ -1,5 +1,0 @@
-class EmptyTodo extends DukeException {
-    public EmptyTodo() {
-        super("☹ OOPS!!! The description of a todo cannot be empty.");
-    }
-}
