+++ src/main/java/EmptyTodoException.java
@@ -1,0 +1,5 @@
+class EmptyTodoException extends DukeException {
+    public EmptyTodoException() {
+        super("☹ OOPS!!! The description of a todo cannot be empty.");
+    }
+}
