+++ src/main/java/InvalidInputException.java
@@ -1,0 +1,5 @@
+class InvalidInputException extends DukeException {
+    public InvalidInputException() {
+        super("☹ OOPS!!! I'm sorry, but I don't know what that means :-(");
+    }
+}
