--- src/main/java/InvalidTask.java
@@ -1,5 +1,0 @@
-class InvalidInput extends DukeException {
-    public InvalidInput() {
-        super("☹ OOPS!!! I'm sorry, but I don't know what that means :-(");
-    }
-}
